fix: make .wire reflect Vorlax incidence angles
Change .wire output to roughly represent what Vorlax "sees". In particular, Vorlax does not shorten the x-dimension of a surface when it is deflected by the input angle. Thus, it is incumbent on the user to account for reduction in chord when surfaces are deflected or twisted. For small angles, the effect can probably be neglected, particularly at deflections at which the flow would remain attached. In any case, keep in mind that the resulting cartoon (i.e., the x-y-z points in the ".wire" file) is only a graphical means of checking size and placement of surfaces, the sense and magnitude of deflections, and the overall accuracy of the input set. The .wire and associated plots are outputs, not inputs. Do not think of the .wire values as a model being placed in a wind tunnel. Vorlax "sees" something slightly different, as described by the .log file, where for instance the deflections are represented by a "slope" associated with the surface, not a physical displacement of the surface.